pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["Wendel <wendelkleberty@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pytest = "^7.4.0"
factory-boy = "^3.3.0"
django-rest-framework = "^0.1.0"
django-extensions = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
